Egen route for minoppgavelista + oppsplitting av de to med egne wrappere for å skille på logikken
