+++ apps/etterlatte-saksbehandling-ui/client/src/App.tsx
@@ -28,6 +28,7 @@ import { isSuccess } from '~shared/api/apiUtils'
 import { isFailureHandler } from '~shared/api/IsFailureHandler'
 import { FlyttJournalpost } from '~components/person/flyttjournalpost/FlyttJournalpost'
 import { FEATURE_TOGGLE_KAN_BRUKE_KLAGE } from '~components/person/KlageListe'
+import { MinOppgaveliste } from '~components/oppgavebenk/MinOppgaveliste'
 
 function App() {
   const innloggetbrukerHentet = useInnloggetSaksbehandler()
@@ -54,8 +55,9 @@ function App() {
             <ErrorBoundary>
               <ConfigContext.Provider value={hentConfigStatus.data}>
                 <Routes>
-                  <Route path="/" element={<ToggleMinOppgaveliste />} />
-                  <Route path="/oppgavebenken" element={<ToggleMinOppgaveliste />} />
+                  <Route path="/" element={<ToggleMinOppgaveliste />}>
+                    <Route path="minoppgaveliste" element={<MinOppgaveliste />} />
+                  </Route>
                   <Route path="/flyttjournalpost" element={<FlyttJournalpost />} />
                   <Route path="/person/:fnr" element={<Person />} />
                   <Route path="/oppgave/:id/*" element={<BehandleJournalfoeringOppgave />} />
