+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/filter/oppgavelistafiltre.ts
@@ -254,3 +254,18 @@ export const initialFilter = (): Filter => {
     fnrSortering: 'ingen',
   }
 }
+
+export const minOppgavelisteFiltre = (): Filter => {
+  return {
+    enhetsFilter: 'visAlle',
+    fristFilter: 'visAlle',
+    saksbehandlerFilter: 'visAlle',
+    ytelseFilter: 'visAlle',
+    oppgavestatusFilter: [OPPGAVESTATUSFILTER.UNDER_BEHANDLING],
+    oppgavetypeFilter: 'visAlle',
+    oppgavekildeFilter: 'visAlle',
+    fnrFilter: '',
+    fristSortering: 'ingen',
+    fnrSortering: 'ingen',
+  }
+}
